+++ packages/analytics/src/server/linkedin-conversions-api.ts
@@ -3,7 +3,7 @@
  * https://learn.microsoft.com/en-us/linkedin/marketing/conversions/conversions-overview?view=li-lms-2025-09
  */
 import { createHash } from 'crypto';
-import { fetch } from '../utils/fetch';
+import { fetch } from '@shware/utils';
 import { getFirst } from '../utils/field';
 import { IGNORE_EVENTS } from './ignore-events';
 import type { TrackEvent, UserProvidedData } from '../track/types';