+++ packages/analytics/src/server/reddit-conversions-api.ts
@@ -1,5 +1,5 @@
+import { fetch } from '@shware/utils';
 import { mapRDTEvent, mapServerStandardEvent } from '../track/rdt';
-import { fetch } from '../utils/fetch';
 import { getFirst } from '../utils/field';
 import { IGNORE_EVENTS } from './ignore-events';
 import type { ServerStandardEvent } from '../track/rdt';